First working proto
dtreematch compares the contents of two directory trees recursively
to check if they are similar. The paths can also be mount points.
This is an attempt at quickly figuring out if the contents of two paths
are not similar.

The match(comparison) happens in this order for an entry(file,dir) in
the path:

  - File pathname check - file availability
  - File type/format check - regular file, directory, block, link, etc.
  - File size check
  - Checksum check

If any of the above check fails for an entry while traversing(the very
first mismatch), without continuing any further, the program bails out
saying 'FAIL' along with the mismatch or program failure reason. If
the contents are similar, exits with a string saying 'MATCH'.

Signed-off-by: six-k <ramsri.hp@gmail.com>